feat(navigation): update navigation
