+++ src/app/navigation/index.ts
@@ -1,3 +1,3 @@
 export * from './home-navigator';
-export * from './root-navigator';
+export * from './welcome-navigator';
 export * from './utils/navigation-service';
